requirements.txt
spotipy
# sqlite3
python-dotenv
requests